rhomb(feature): GPIO test utility added as part of buildroot
This commit adds GPIO hardware test utility. This is userspace utility to
modify GPIO value and direction. It is very simple to operate. Help is
provided as part of utiliy itself. In the target filesystem utility will
be availalbe at "/opt/rhomb/hw_tests" directory. It is generic for all the
future utilites also.

Signed-off-by: Android Linux Media <android.linuxmedia@gmail.com>